feat(react): export type InferBundleRoom
