+++ packages/react/src/index.ts
@@ -2,7 +2,8 @@ export { createBundle } from "./createBundle";
 export type {
     CreateBundle,
     EventProxy,
+    InferBundleRoom,
     PluvProviderProps,
     PluvRoomProviderProps,
     SubscriptionHookOptions,
-} from "./createBundle";
+} from "./types";
